refactor: 패키지 이동
session 패지를 catalina 하위로 이동
