+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -4,8 +4,8 @@ import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.io.IOException;
 import java.util.UUID;
-import org.apache.catalina.SessionManager;
-import org.apache.coyote.http11.Session;
+import org.apache.catalina.session.SessionManager;
+import org.apache.catalina.session.Session;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
